third_party/blink/web_tests/external/wpt/wasm/jsapi/module/moduleSource.tentative.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] AbstractModuleSource intrinsic
  assert_equals: expected "AbstractModuleSource" but got ""
[FAIL] AbstractModuleSourceProto intrinsic
  assert_not_equals: got disallowed value object "[object Object]"
[FAIL] AbstractModuleSourceProto toStringTag brand check
  Cannot convert undefined or null to object
Harness: the test ran to completion.
